remove comments, pull out utils to discrete file
